+++ src/dashboard/dash-board-about.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import DashBoardMenu from './DashBoardMenu';
+import DashBoardMenu from './dash-board-menu';
