+++ src/dashboard/dash-board-breadcrumb-bright.js
