+++ src/dashboard/dash-board-title-bright.js
