+++ src/front/reg-and-login-button.js
