+++ src/item-for-voting.js
